# Don't coerce sparse data to non-sparse during predict()