This bug was first reported here: https://community.rstudio.com/t/predict-not-working-with-ranger-model-when-using-sparse-data/163352/7

When you call `predict()` you eventually get to run `prepare_data()`, which didn't know about the `alllow_sparse_x` encoding so it would try to turn it into matrices/data.frames. This was bad for 2 reasons. First we are losing some performance. Secondly, in the case of the ranger method, it wants the data as a data.frame which a sparse data can't be turned into, yielding the error seen below.

# Main

``` r
library(parsnip)

data(agaricus.train, package = 'xgboost')
train <- agaricus.train

rf_model <- parsnip::rand_forest(trees = 100) %>% 
  set_engine("ranger") %>% 
  set_mode("classification")

rf_fit <- fit_xy(rf_model, x = train$data, y = factor(train$label))

predict(rf_fit, train$data)
#> Error in as.data.frame.default(new_data): cannot coerce class 'structure("dgCMatrix", package = "Matrix")' to a data.frame
```

# This PR

``` r
library(parsnip)

data(agaricus.train, package = 'xgboost')
train <- agaricus.train

rf_model <- parsnip::rand_forest(trees = 100) %>% 
  set_engine("ranger") %>% 
  set_mode("classification")

rf_fit <- fit_xy(rf_model, x = train$data, y = factor(train$label))

predict(rf_fit, train$data)
#> # A tibble: 6,513 × 1
#>    .pred_class
#>    <fct>      
#>  1 1          
#>  2 0          
#>  3 0          
#>  4 1          
#>  5 0          
#>  6 0          
#>  7 0          
#>  8 1          
#>  9 0          
#> 10 0          
#> # ℹ 6,503 more rows
```