@EmilHvitfeldt 2023-04-12
Small speedup benchmark. (this is likely to vary majorly for different data, but it should almost always be in the right direction

<details>
  <summary>Code used for the following reprexes</summary>

```r
library(parsnip)

data(agaricus.train, package = 'xgboost')
train <- agaricus.train

xg_model <- parsnip::boost_tree(trees = 100) %>% 
  set_engine("xgboost") %>% 
  set_mode("classification")

xg_fit <- fit_xy(xg_model, x = train$data, y = factor(train$label))
```
</details>

# Main

``` r
bench::mark(
  old = predict(xg_fit, train$data)
)
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old          8.96ms   9.55ms      105.     7.2MB     60.4
```

# This PR

``` r
bench::mark(
  new = predict(xg_fit, train$data)
)
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 new          6.24ms   6.65ms      150.    1.13MB     4.17
```

<sup>Created on 2023-04-12 with [reprex v2.0.2](https://reprex.tidyverse.org)</sup>